Upgrade all length parameters of RichString class to size_t

+++ linux/GPUMeter.c
@@ -126,10 +126,10 @@ static void GPUMeter_display(const Object* cast, RichString* out) {
 
    RichString_writeAscii(out, CRT_colors[METER_TEXT], ":");
    written = xSnprintf(buffer, sizeof(buffer), "%4.1f", totalUsage);
-   RichString_appendnAscii(out, CRT_colors[METER_VALUE], buffer, written);
+   RichString_appendnAscii(out, CRT_colors[METER_VALUE], buffer, (unsigned int)written);
    RichString_appendAscii(out, CRT_colors[METER_TEXT], "%(");
    written = humanTimeUnit(buffer, sizeof(buffer), totalGPUTimeDiff);
-   RichString_appendnAscii(out, CRT_colors[METER_VALUE], buffer, written);
+   RichString_appendnAscii(out, CRT_colors[METER_VALUE], buffer, (unsigned int)written);
    RichString_appendAscii(out, CRT_colors[METER_TEXT], ")");
 
    for (i = 0; i < ARRAYSIZE(GPUMeter_engineData); i++) {
@@ -143,10 +143,10 @@ static void GPUMeter_display(const Object* cast, RichString* out) {
          written = xSnprintf(buffer, sizeof(buffer), "%4.1f", this->values[i]);
       else
          written = xSnprintf(buffer, sizeof(buffer), " N/A");
-      RichString_appendnAscii(out, CRT_colors[METER_VALUE], buffer, written);
+      RichString_appendnAscii(out, CRT_colors[METER_VALUE], buffer, (unsigned int)written);
       RichString_appendAscii(out, CRT_colors[METER_TEXT], "%(");
       written = humanTimeUnit(buffer, sizeof(buffer), GPUMeter_engineData[i].timeDiff);
-      RichString_appendnAscii(out, CRT_colors[METER_VALUE], buffer, written);
+      RichString_appendnAscii(out, CRT_colors[METER_VALUE], buffer, (unsigned int)written);
       RichString_appendAscii(out, CRT_colors[METER_TEXT], ")");
    }
 }
