+++ pcp/PCPDynamicMeter.c
@@ -461,7 +461,7 @@ void PCPDynamicMeter_display(PCPDynamicMeter* this, ATTR_UNUSED const Meter* met
       }
 
       if (len) {
-         RichString_appendnAscii(out, CRT_colors[metric->color], buffer, len);
+         RichString_appendnAscii(out, CRT_colors[metric->color], buffer, (unsigned int)len);
          if (metric->suffix)
             RichString_appendAscii(out, CRT_colors[METER_TEXT], metric->suffix);
       }
